Factor a container argument generator out of the wdl code.
The wdl code now only handles WDL specific stuff.
