+++ cdmtaskservice/jaws/wdl.py
@@ -5,23 +5,15 @@ A builder of Workflow Definition Language documents for the purposes of running 
 # TODO TEST parse output with miniwdl or something for syntax checking
 # TODO TEST write manual test for testing with JAWS
 
-import os
 from pathlib import Path
 import math
 import shlex
 from typing import NamedTuple, Any
 
+from cdmtaskservice.argument_generator import ArgumentGenerator
 from cdmtaskservice.arg_checkers import not_falsy as _not_falsy
-from cdmtaskservice.models import (
-    Job,
-    S3FileWithDataID,
-    Parameter,
-    ParameterType,
-    InputFilesFormat,
-    PATH_REGEX_COMPILED,
-)
-from cdmtaskservice.input_file_locations import determine_file_locations
 from cdmtaskservice.jaws.constants import OUTPUT_FILES, OUTPUT_DIR, STDOUTS, STDERRS
+from cdmtaskservice.models import Job, S3FileWithDataID, PATH_REGEX_COMPILED
 
 
 _WDL_VERSION = "1.0"  # Cromwell, and therefore JAWS, only supports 1.0
@@ -61,55 +53,37 @@ def generate_wdl(
         This will be mounted into the container at the refdata mount point specified in the job.
     """
     # It'd be nice if there were a programmatic WDL writer but I haven't been able to find one
-    # This fn is a little long but not too hard to read yet I think
     # If the manifest file name / path changes that'll break the JAWS cache, need to think
     # about that
     # How often will jobs run with identical manifest files though? Maybe MD5 based caching?
     _not_falsy(job, "job")
     _not_falsy(file_mapping, "file_mapping")
+    arggen = ArgumentGenerator(job, manifest_file_list=manifest_file_list)
     if relative_refdata_path and not job.get_refdata_mount_point():
         raise ValueError(
             "If a refdata path is supplied, a mount point for the job must be supplied"
         )
-    if not job.job_input.inputs_are_S3File():
-        raise ValueError("input files must be S3 files with a checksum")
-    param = job.job_input.params.get_file_parameter()
-    if param and param.type is ParameterType.MANIFEST_FILE and (
-        not manifest_file_list or job.job_input.num_containers != len(manifest_file_list)
-    ):
-        raise ValueError(
-            "If a manifest file is specified in the job parameters, manifest_file_list "
-            + "is required and its length must match the number of containers for the job"
-        )
-    job_files = job.job_input.get_files_per_container()
-    if manifest_file_list:
-        for m in manifest_file_list:
-            if not PATH_REGEX_COMPILED.fullmatch(str(m)):
-                raise ValueError(f"Disallowed manifest path: {m}")
     for s3o, localpath in file_mapping.items():
         if not PATH_REGEX_COMPILED.fullmatch(str(localpath)):
             raise ValueError(f"Disallowed local path for S3 object {s3o}: {localpath}")
     workflow_name = job.image.name.translate(_IMAGE_TRANS_CHARS)
-    file_to_rel_path = determine_file_locations(job.job_input)
     input_files = []
     relpaths = []
     environment = []
     cmdlines = []
-    mfl = [None] * job.job_input.num_containers if not manifest_file_list else manifest_file_list
-    for i, (files, manifest) in enumerate(zip(job_files, mfl)):
+    for container_number in range(job.job_input.num_containers):
         ins = []
         rels = []
-        for f in files:
+        arginfo = arggen.get_container_arguments(container_number)
+        for f in arginfo.files:
             if f not in file_mapping:
                 raise ValueError(f"file_mapping missing {f}")
             ins.append(shlex.quote(str(file_mapping[f])))
-            rels.append(shlex.quote(str(file_to_rel_path[f])))
+            rels.append(shlex.quote(str(arginfo.files[f])))
         input_files.append(ins)
         relpaths.append(rels)
-        cmd = [shlex.quote(c) for c in job.image.entrypoint]
-        cmd.extend(_process_pos_args(job, i, files, file_to_rel_path, manifest))
-        cmdlines.append(cmd)
-        environment.append(_process_environment(job, i, files, file_to_rel_path, manifest))
+        cmdlines.append(arginfo.args)
+        environment.append(arginfo.env)
     input_json = {
         f"{workflow_name}.input_files_list": input_files,
         f"{workflow_name}.file_locs_list": relpaths,
@@ -233,153 +207,3 @@ task run_container {{
   }}
 }}
 """
-
-
-def _process_pos_args(
-    job: Job,
-    container_num: int,
-    files: list[S3FileWithDataID],
-    file_to_rel_path: dict[S3FileWithDataID, Path],
-    manifest: Path | None,
-) -> list[str]:
-    cmd = []
-    if job.job_input.params.args:
-        for p in job.job_input.params.args:
-            cmd.extend(_process_parameter(
-                p, job, container_num, files, file_to_rel_path, manifest, as_list=True
-            ))
-    return cmd
-
-
-def _process_environment(
-    job: Job,
-    container_num: int,
-    files: list[S3FileWithDataID],
-    file_to_rel_path: dict[S3FileWithDataID, Path],
-    manifest: Path | None,
-) -> list[str]:
-    env = []
-    if job.job_input.params.environment:
-        for envkey, enval in job.job_input.params.environment.items():
-            enval = _process_parameter(
-                enval, job, container_num, files, file_to_rel_path, manifest, no_flag=True
-            )
-            env.append(f'{envkey}={enval}')
-    return env
-
-
-def _process_parameter(
-    param: str | Parameter,
-    job: Job,
-    container_num: int,
-    files: list[S3FileWithDataID],
-    file_to_rel_path: dict[S3FileWithDataID, Path],
-    manifest: Path | None,
-    as_list: bool = False,
-    no_flag: bool = False,
-) -> str | list[str]:
-    if isinstance(param, Parameter):
-        flag = None if no_flag else param.get_flag()
-        match param.type:
-            case ParameterType.INPUT_FILES:
-                param = _join_files(
-                    files, param.input_files_format, job, file_to_rel_path, flag
-                )
-            case ParameterType.MANIFEST_FILE:  # implies manifest file is not None
-                param = _handle_manifest(job, manifest, flag)
-            case ParameterType.CONTAINTER_NUMBER:
-                param = _handle_container_num(param, container_num, flag)
-            case _:
-                # should be impossible but make code future proof
-                raise ValueError(f"Unexpected parameter type: {_}")
-    else:
-        param = shlex.quote(param)
-    return [param] if as_list and not isinstance(param, list) else param
-
-
-def _handle_container_num(p: Parameter, container_num: int, flag: str | None) -> str | list[str]:
-    # similar to the function below
-    pre = p.container_num_prefix if p.container_num_prefix else ""
-    suf = p.container_num_suffix if p.container_num_suffix else ""
-    cn = f"{pre}{container_num}{suf}"
-    if flag:
-        if flag.endswith("="):
-            # not sure if this will work, but since Job flags are shell safe it's
-            # impossible to test
-            param = shlex.quote(flag) + cn
-        else:
-            param = [shlex.quote(flag), cn]
-    else:
-        param = shlex.quote(cn)
-    return param
-
-
-def _handle_manifest(job: Job, manifest: Path, flag: str | None) -> str | list[str]:
-    pth = os.path.join(job.job_input.params.input_mount_point, manifest.name)
-    # This is the same as the comma separated list case below... not sure if using a common fn
-    # makes sense
-    if flag:
-        if flag.endswith("="):
-            param = shlex.quote(flag + pth)
-        else:
-            param = [shlex.quote(flag), shlex.quote(pth)]
-    else:
-        param = shlex.quote(pth)
-    return param
-
-
-# this is a bit on the complex side...
-def _join_files(
-    files: list[S3FileWithDataID],
-    format_: InputFilesFormat,
-    job: Job,
-    file_to_rel_path: dict[S3FileWithDataID, Path],
-    flag: str | None,
-) -> str | list[str]:
-    imp = job.job_input.params.input_mount_point
-    match format_:
-        case InputFilesFormat.COMMA_SEPARATED_LIST:
-            fileret = ",".join([_join_path(imp, file_to_rel_path, f) for f in files])
-            if flag:
-                if flag.endswith("="):
-                    fileret = shlex.quote(flag + fileret)
-                else:
-                    fileret = [shlex.quote(flag), shlex.quote(fileret)]
-            else:
-                fileret = shlex.quote(fileret)
-        case InputFilesFormat.SPACE_SEPARATED_LIST:
-            if flag:
-                if flag.endswith("="):
-                    # this case is a little weird
-                    fileret = [
-                        shlex.quote(flag + _join_path(imp, file_to_rel_path, files[0]))
-                    ] + [_join_path(imp, file_to_rel_path, f, quote=True) for f in files[1:]]
-                else: 
-                    fileret = [shlex.quote(flag)] + [
-                        _join_path(imp, file_to_rel_path, f, quote=True) for f in files
-                    ]
-            else:
-                fileret = [_join_path(imp, file_to_rel_path, f, quote=True) for f in files]
-        case InputFilesFormat.REPEAT_PARAMETER:  # implies flag
-            fileret = []
-            if flag.endswith("="):
-                for f in files:
-                    fileret.append(shlex.quote(flag + _join_path(imp, file_to_rel_path, f)))
-            else:
-                flag = shlex.quote(flag)
-                for f in files:
-                    fileret.extend((flag, _join_path(imp, file_to_rel_path, f, quote=True)))
-        case _:
-            # should be impossible but make code future proof
-            raise ValueError(f"Unexpected input files format: {_}")
-    return fileret
-
-
-def _join_path(
-        imp: str,
-        file_to_rel_path: dict[S3FileWithDataID, Path],
-        f: S3FileWithDataID,
-        quote: bool = False
-) -> str:
-    p = os.path.join(imp, file_to_rel_path[f])
-    return shlex.quote(p) if quote else p
