+++ test/argument_generator_test.py
@@ -1,0 +1,10 @@
+# TODO TEST add tests
+
+# Note that the jaws/wdl_test.py tests cover this well enough for now, although eventually unit
+# tests would be good.
+
+from cdmtaskservice import argument_generator  # @UnusedImport
+
+
+def test_noop():
+    pass
